+++ src/marksapp/models.py
@@ -6,6 +6,7 @@ from django.core.validators import validate_slug
 
 class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
+    email = models.CharField(max_length=128, blank=True)
 
     PUBLIC = 'PB'
     PRIVATE = 'PV'
@@ -14,8 +15,12 @@ class Profile(models.Model):
         (PRIVATE, 'Private',)
     ]
     visibility = models.CharField(choices=visibility_choices,
-                                  max_length=2)
-    email = models.CharField(max_length=128, blank=True)
+                                  max_length=2,
+                                  default=PUBLIC)
+
+    def __str__(self):
+        return "{} | {}".format(self.user.username,
+                                self.visibility)
 
 class Tag(models.Model):
     name = models.SlugField(max_length=50, unique=True)